third_party/blink/web_tests/fast/canvas-api/winding-enumeration.html

<!DOCTYPE html>
<title>Test CanvasWindingRule enumeration</title>

<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>

<script>
test(function() {
    var canvas = document.createElement("canvas");
    var context = canvas.getContext("2d");

    context.fill();
    context.fill('nonzero');
    assert_throws_js(TypeError, function() { context.fill('randomstring'); });

    context.clip();
    context.clip('nonzero');
    assert_throws_js(TypeError, function() { context.clip('randomstring'); });

    context.isPointInPath(0, 0);
    context.isPointInPath(0, 0, 'nonzero');
    assert_throws_js(TypeError, function() { context.isPointInPath(0, 0, 'randomstring'); });
});
</script>
